Add unit tests for word trends and word trend speeches router endpoints
- Implemented comprehensive unit tests for the word trends router, covering endpoints for retrieving word trends results, submitting queries, checking ticket statuses, and downloading speech data.
- Removed unused test for topics in integration tests.
- Cleaned up assertions in ticket validation tests for better readability and consistency.
