+++ tests/api_swedeb/api/endpoints/test_ngrams_router.py
@@ -1,0 +1,7 @@
+"""Unit tests for ngrams router endpoints.
+
+Handler-level unit tests for the async ticket endpoints (submit_ngrams_query,
+get_ngrams_ticket_status, get_ngrams_ticket_page, prepare_ngrams_archive) are
+not yet in the suite. See tests/integration/ for ngrams coverage via the HTTP
+layer.
+"""
